Rage no longer gives its users a temporary strength buff, but instead increases its overall damage output by a certain percentage.
- Rage also increases the amount of damage taken by its users for a short amount of time.
- Fixed nametag color for the soul obliterator
- Modified priorities of entitydamagebyentity listeners to allow changes in incoming and dealt damage to be recognized by our damage display listener.

+++ AGCraftPlugins/src/com/joojet/plugins/mobs/skills/passive/interfaces/PassiveAttack.java
@@ -15,7 +15,7 @@ public interface PassiveAttack
 	 *  @param target The LivingEntity that is victim of the damage event
 	 *  @param damagerEquipment The MobEquipment class belonging to the damager
 	 *  @return Bonus damage dealt to the target (not incliding the original damage)*/
-	public int modifyOutgoingDamageEvent (double damage, LivingEntity damager, LivingEntity target, MobEquipment damagerEquipment);
+	public double modifyOutgoingDamageEvent (double damage, LivingEntity damager, LivingEntity target, MobEquipment damagerEquipment);
 	
 	/** Allows custom behavior to be applied onto any damage taken by the skill caster with this skill.
 	 *  @param damage Original damage dealt to the target
@@ -23,5 +23,5 @@ public interface PassiveAttack
 	 *  @param target The LivingEntity that is victim of the damage event
 	 *  @param targetEquipment The MobEquipment class belonging to the target
 	 *  @return Bonus damage taken by the target (not incliding the original damage)*/
-	public int modifyIncomingDamageEvent (double damage, LivingEntity damager, LivingEntity target, MobEquipment targetEquipment);
+	public double modifyIncomingDamageEvent (double damage, LivingEntity damager, LivingEntity target, MobEquipment targetEquipment);
 }
